fix: reduce connect auth startup diff
