--- apps/web/src/components/connect/T3ConnectSidebarSignIn.configured.tsx
@@ -1,52 +1,0 @@
-import { UserButton, useAuth, useClerk } from "@clerk/react";
-import { LogInIcon, SmartphoneIcon } from "lucide-react";
-
-import { SidebarMenu, SidebarMenuButton, SidebarMenuItem } from "../ui/sidebar";
-import { MobileClientsUserProfilePage } from "./MobileClientsUserProfilePage";
-
-export function ConfiguredT3ConnectSidebarAvatar() {
-  const { isLoaded, isSignedIn } = useAuth();
-
-  if (!isLoaded || !isSignedIn) return null;
-
-  return (
-    <UserButton
-      appearance={{
-        elements: {
-          avatarBox: "size-7",
-          userButtonTrigger: "rounded-lg p-1 hover:bg-sidebar-accent",
-        },
-      }}
-    >
-      <UserButton.UserProfilePage
-        label="Mobile clients"
-        labelIcon={<SmartphoneIcon className="size-4" />}
-        url="mobile-clients"
-      >
-        <MobileClientsUserProfilePage />
-      </UserButton.UserProfilePage>
-    </UserButton>
-  );
-}
-
-export function ConfiguredT3ConnectSidebarSignIn() {
-  const { isLoaded, isSignedIn } = useAuth();
-  const clerk = useClerk();
-
-  if (!isLoaded || isSignedIn) return null;
-
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <SidebarMenuButton
-          size="sm"
-          className="gap-2 px-2 py-2 text-xs text-muted-foreground hover:bg-accent hover:text-foreground"
-          onClick={() => clerk.openWaitlist()}
-        >
-          <LogInIcon className="size-4" />
-          <span>Sign in to T3 Connect</span>
-        </SidebarMenuButton>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
-}
